requirements.txt
torch>=2.0
torchvision
torchmetrics
pillow
numpy
facenet-pytorch
